revert r11934, causes javac warning "Compound assignments to bytes, shorts, chars, and floats hide dangerous casts"
git-svn-id: http://josm.openstreetmap.de/svn/trunk@11940 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/osm/AbstractPrimitive.java
@@ -343,7 +343,7 @@ public abstract class AbstractPrimitive implements IPrimitive {
         if (value) {
             flags |= flag;
         } else {
-            flags &= ~flag;
+            flags &= (short) ~flag;
         }
     }
